all but test_ui_profile_delete_cbv
